pyproject.toml
[tool.poetry]
name = "burlabet"
version = "0.1.0"
description = ""
authors = ["kevindotklein <kevindotklein@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
beautifulsoup4 = "^4.13.4"
python-dotenv = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
burlabet = "main:main"